third_party/WebKit/LayoutTests/external/wpt/pointerevents/pointerevent_click_during_capture-manual-expected.txt

This is a testharness.js-based test.
FAIL mouse click target during capture assert_equals: An element should only receive click when it is the first common ancestor of pointerdown and pointerup targets expected "pointerdown@green,gotpointercapture@blue,pointerup@blue,click@grey,lostpointercapture@blue" but got "pointerdown@green,gotpointercapture@blue,pointerup@blue,click@green,lostpointercapture@blue"
PASS PointerEvent Automation 
PASS Click target when pointerup/down targeted at the same element with no capture 
PASS Click target when pointerup/down targeted at different elements with no capture 
Harness: the test ran to completion.
